Puzzle 10
I could see a recursive solution to part 2 however the puzzle text
hinted that this wouldn't be feasible. Spent a lot of time looking
at my diff Vec trying to see a shortcut. Eventually came up with
this solution after playing around with pen and paper. Part two of
today is easily the toughest so far for me.