Update puzzle 11.
This was running in ~600ms. Switched to a Vec instead of a HashMap
and this was an order of magnitude faster. Now around ~50-70ms.
Still seems a bit slow. One thought, which others have done, is to
use a custom Grid struct that stores the data in a single Vec
instead of nested Vecs. Could then add a get() method that handles
the out of bounds checks. Avoiding the double array lookups might
help. Perf and flamegraphs showed the count_occupied_neighbours()
function was very hot.